feature: relocate common funcs in one file
